+++ src/java/abfab3d/grid/BitIntervals.java
@@ -61,7 +61,7 @@ public class BitIntervals implements RowOfInt, Serializable {
     /**
 
      */
-    public long get(int x){
+    public int get(int x){
 
         //find interval which contains x 
         //
